updated settings handling for q axes

+++ include/core/api/api_settings.h
@@ -44,7 +44,7 @@ extern "C" API void set_hist_settings(
     double qmax,
     bool weighted_bins,
     double bin_width,
-    double bin_count,
+    unsigned int bin_count,
     int* status
 );
